% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessTreeNumberConsistency.R
\name{assessTreeNumberConsistency}
\alias{assessTreeNumberConsistency}
\title{Analyze Regeneration Proportion and Tree DBH Growth}
\usage{
assessTreeNumberConsistency(plots, max_assumed_growth_rate = 1)
}
\arguments{
\item{plots}{A list containing at least a data frame \code{PSPmeasure} with tree measurements and \code{PSPplot} with measurement years.}

\item{max_assumed_growth_rate}{Maximum plausible DBH growth per year for
regenerating trees (cm/year, default = 1).}
}
\value{
A list with:
\describe{
\item{problematicMeasurements}{MeasurementIDs with 100\% regen at some point in plot history
If the problematic measurement is the last measurement, then prior measurements are not included.
If the problematic measurement is NOT the first, then all measurements are included}
\item{problematic trees}{trees with problematic DBH (ie growing faster than max_assumed_growth_rate) and other
variables: elapsedTime (years between measurements), DBHdiffFromMax (the difference between the actual growth rate
and the expected maximum growth rate), and DBHdiffFromMax_perYear (the latter divided by the former))}
}
}
\description{
\enumerate{
\item Identifies plots where all trees identified, and their subsequent measurements. These are of interest
as the stand age is unlikely to be correct
\item calculates elapsed time between successive measurements and flags trees with
biologically implausible DBH growth based on the max_assumed_growth_rate param
}
}
