% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFIPSP.R
\name{dataPurification_NFIPSP}
\alias{dataPurification_NFIPSP}
\title{Standardize and Treat the NFI PSP Data}
\usage{
dataPurification_NFIPSP(
  NFIdata,
  codesToExclude = "IB",
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{NFIdata}{A list containing NFI tree, plot, and location data.tables.}

\item{codesToExclude}{Vector of damage agent codes. Measurements with these codes will be removed.}

\item{excludeAllObs}{Logical. If \code{TRUE}, removing observations of individual trees due to damage codes
will also remove all prior and future observations of that tree.}

\item{sppEquiv}{A table providing species name equivalencies between the original PSP species names
and the final standardized naming format. Default is \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{Character string. The column in \code{sppEquiv} that contains the standardized species names.
Default is \code{"Latin_full"}.}
}
\value{
A list containing standardized \code{plotData} and \code{treeData} as \code{data.table}s.
}
\description{
This function cleans and standardizes NFI PSP data, including tree, plot, and location data.
Species names can be standardized using a species equivalency table.
}
