% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntarioPSP.R
\name{dataPurification_ONPSP}
\alias{dataPurification_ONPSP}
\title{standardize and treat the Ontario PSP data}
\usage{
dataPurification_ONPSP(
  ONPSPlist,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{ONPSPlist}{list of relevant plots}

\item{sppEquiv}{sdfsd}

\item{sppEquivCol}{sdfd}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the Ontario PSP data
}
