% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuebecPSP.R
\name{dataPurification_QCPSP}
\alias{dataPurification_QCPSP}
\title{Standardize and Treat the Quebec PSP Data}
\usage{
dataPurification_QCPSP(
  QuebecPSP,
  codesToExclude = NULL,
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{QuebecPSP}{A list of \code{data.table}s containing raw PSP data obtained via \code{prepInputsQCPSP}.}

\item{codesToExclude}{Character vector of damage or pest disturbance codes to exclude from measurements.
Currently placeholder, update if applicable.}

\item{excludeAllObs}{Logical. If \code{TRUE} and \code{codesToExclude} is not \code{NULL}, all prior and future
observations of trees with these codes will be removed.}

\item{sppEquiv}{A table providing species name equivalencies between the original PSP species names
and the final standardized naming format. Default is \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{Character string. The column in \code{sppEquiv} that contains the standardized species names.
Default is \code{"Latin_full"}.}
}
\value{
A list containing standardized \code{plotData} and \code{treeData} as \code{data.table}s.
}
\description{
This function cleans and standardizes the Quebec PSP data, including tree and plot data.
Species names can be standardized using a species equivalency table.
}
