% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaskatchewanPSP.R
\name{dataPurification_SKPSP}
\alias{dataPurification_SKPSP}
\title{standardize and treat the Saskatchewan PSP data}
\usage{
dataPurification_SKPSP(
  SADataRaw,
  plotHeaderRaw,
  measureHeaderRaw,
  treeDataRaw,
  codesToExclude = NULL,
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{SADataRaw}{the tree measurement csv}

\item{plotHeaderRaw}{the plot header data}

\item{measureHeaderRaw}{the measurement header raw}

\item{treeDataRaw}{tree data}

\item{codesToExclude}{damage agent codes used to filter tree data.
Natural or Undetermined = 1, Disease = 2, Insect = 3, Human 4,
Wind = 5, Snow =  6, Other Trees = 7, Hail or Ice Storm = 8.
Measurements with these codes will be removed}

\item{excludeAllObs}{if removing observations of individual trees due to damage codes,
remove all prior and future observations if \code{TRUE}.}

\item{sppEquiv}{sdfsd}

\item{sppEquivCol}{sdfd}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the Saskatchewan PSP data
}
