% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R
\name{detect_DBH_outliers}
\alias{detect_DBH_outliers}
\title{#-----------------------------------------------------------   detect_DBH_outliers function()  ----------------------------------------------
Detect and Flag DBH Outliers}
\usage{
detect_DBH_outliers(Trees, maxDBHrealistic = 400)
}
\arguments{
\item{Trees}{A \code{data.table} with columns: "DBH", "OrigPlotID1", "TreeNumber", and "MeasureYear".}

\item{maxDBHrealistic}{threshold for realistic DBH that is not an error}
}
\value{
A list containing:
\describe{
\item{Trees}{The full data.table with \code{DBH_flag}, \code{DBH_growth}, and \code{Growth_flag}.}
\item{Sapling}{Subset flagged as "Sapling".}
\item{Impossible}{Subset flagged as "Impossible".}
\item{Exceptional}{Subset flagged as "Exceptional".}
\item{VerifyData}{Subset flagged as "VerifyData".}
\item{OrigPlotID1s}{Unique OrigPlotID1 identifiers.}
}
}
\description{
Flags implausible DBH values (static outliers) and unusual year-over-year DBH changes (dynamic outliers)
to help identify data issues in repeated tree measurements.
}
