% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoCleanPSP.R
\name{geoCleanPSP}
\alias{geoCleanPSP}
\title{convert plot location data to long/lat and output in a sf object}
\usage{
geoCleanPSP(Locations)
}
\arguments{
\item{Locations}{the compiled plot data}
}
\value{
an sf object with \code{OrigPlotID1} column
}
\description{
convert plot location data to long/lat and output in a sf object
}
