% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regen_proportion.R
\name{plot_regen_proportion}
\alias{plot_regen_proportion}
\title{Analyze Regeneration Proportion and Tree DBH Growth}
\usage{
plot_regen_proportion(plots, max_assumed_growth_rate = 1)
}
\arguments{
\item{plots}{A list containing at least a data frame \code{PSPmeasure} with tree measurements and \code{PSPplot} with measurement years.}

\item{max_assumed_growth_rate}{Maximum plausible DBH growth per year for
regenerating trees (cm/year, default = 1).}
}
\value{
A list with:
\describe{
\item{problematicMeasurements}{MeasurementIDs with 100\% regen at some point in plot history
If the problematic measurement is the last measurement, then prior measurements are not included.
If the problematic measurement is NOT the first, then all measurements are included}
\item{problematic trees}{trees with problematic DBH (ie growing faster than max_assumed_growth_rate)}
}
}
\description{
Computes the proportion of regeneration trees (\code{propStatus}) per MeasureID, Plot, and Source;
identifies high-regeneration MeasureIDs; summarizes DBH statistics for high-regeneration plots;
calculates elapsed time between successive measurements; and flags trees with biologically implausible DBH growth.
}
