% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlbertaPSP.R
\name{prepInputsAlbertaPSP}
\alias{prepInputsAlbertaPSP}
\title{source the Alberta PSP data}
\usage{
prepInputsAlbertaPSP(dPath)
}
\arguments{
\item{dPath}{passed to prepInputs destinationPath}
}
\value{
a list of Alberta PSP data.tables
}
\description{
source the Alberta PSP data
}
