% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPSP.R
\name{prepInputsBCPSP}
\alias{prepInputsBCPSP}
\title{Source the BC PSP data}
\usage{
prepInputsBCPSP(dPath)
}
\arguments{
\item{dPath}{passed to `destinationPath` in `prepInputs()`}
}
\value{
a list of BC PSP objects
}
\description{
Source the BC PSP data
}
