% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewBrunswickPSP.R
\name{prepInputsNBPSP}
\alias{prepInputsNBPSP}
\title{retrieve the New Brunswick PSP raw data}
\usage{
prepInputsNBPSP(dPath)
}
\arguments{
\item{dPath}{data directory for raw data}
}
\value{
a list of plot, tree, measurement, and location data.tables after exporting mdb to csv txt
}
\description{
retrieve the New Brunswick PSP raw data
}
