% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFIPSP.R
\name{prepInputsNFIPSP}
\alias{prepInputsNFIPSP}
\title{source the NFI PSP data}
\usage{
prepInputsNFIPSP(dPath)
}
\arguments{
\item{dPath}{passed to prepInputs destinationPath}
}
\value{
a list of NFI PSP data.tables
}
\description{
source the NFI PSP data
}
