% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntarioPSP.R
\name{prepInputsOntarioPSP}
\alias{prepInputsOntarioPSP}
\title{retrieve preprocessed Ontario PSP and PGP data}
\usage{
prepInputsOntarioPSP(dPath, ...)
}
\arguments{
\item{dPath}{the Access database for PSP and PGP plots}

\item{...}{additional args passed to prepInputs}
}
\value{
a list of plot and tree data.tables
}
\description{
retrieve preprocessed Ontario PSP and PGP data
}
