% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuebecPSP.R
\name{prepInputsQCPSP}
\alias{prepInputsQCPSP}
\title{retrieve the Quebec PSP raw data}
\usage{
prepInputsQCPSP(dPath)
}
\arguments{
\item{dPath}{data directory for raw data}
}
\value{
a list of plot, tree, measurement, and location data.tables after exporting mdb to csv txt
}
\description{
retrieve the Quebec PSP raw data
}
