% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaskatchewanTSP.R
\name{prepInputsSaskatchwanTSP}
\alias{prepInputsSaskatchwanTSP}
\title{source the Saskatchewan PSP data}
\usage{
prepInputsSaskatchwanTSP(dPath)
}
\arguments{
\item{dPath}{passed to prepInputs destinationPath}
}
\value{
a list of Saskatchewan PSP data.tables
}
\description{
source the Saskatchewan PSP data
}
