% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateNAinputs.R
\name{prepPSP_climateNA}
\alias{prepPSP_climateNA}
\title{standardize and treat the BC PSP data}
\usage{
prepPSP_climateNA(dPath, filename2, PSPplot, PSPgis)
}
\arguments{
\item{dPath}{directory to download elevation data}

\item{filename2}{the full filename of the output file}

\item{PSPplot}{the standardized PSP plot attribute data}

\item{PSPgis}{the standardized plot location sf object}
}
\value{
a text file
}
\description{
standardize and treat the BC PSP data
}
