% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSpeciesNames.R
\name{standardizeSpeciesNames}
\alias{standardizeSpeciesNames}
\title{standardize the species names in a given PSP measurement data.table}
\usage{
standardizeSpeciesNames(speciesTable, forestInventorySource)
}
\arguments{
\item{speciesTable}{the tree measurement csv}

\item{forestInventorySource}{one of MBTSP, ABPSP, BCPSP, SKPSP, NWTTSP, NFIPSP.}
}
\value{
the species table with standardized species names
}
\description{
standardize the species names in a given PSP measurement data.table
}
