% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R
\name{classify_tree_status}
\alias{classify_tree_status}
\title{#--------------------------------------------------------------------   Classify tree status function() ---------------------------------------------------------------------}
\usage{
classify_tree_status(Trees)
}
\arguments{
\item{Trees}{A \code{data.table}  of tree records across multiple measurement years.}
}
\value{
A list containing:
\describe{
\item{Trees}{The full dataset with status classification.}
\item{Regeneration}{Trees first appearing after plot monitoring began.}
\item{Last_Measurement}{Trees last measured before monitoring ended.}
\item{Alive}{Trees present consistently or throughout.}
\item{OrigPlotID1s}{Unique OrigPlotID1 identifiers.}
}
}
\description{
Assigns status to each tree based on its observation history. Detects new regeneration, lost trees, and survivors.
}
\details{
Classify Tree Status Based on Measurement History
}
