% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlbertaPSP.R
\name{dataPurification_ABPSP}
\alias{dataPurification_ABPSP}
\title{Standardize and Treat the Alberta PSP Data}
\usage{
dataPurification_ABPSP(
  treeMeasure,
  plotMeasure,
  tree,
  plot,
  codesToExclude = 3,
  excludeAllObs = TRUE,
  areaDiffThresh = 0.95,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{treeMeasure}{A \code{data.frame} or \code{data.table} containing tree measurement data.}

\item{plotMeasure}{A \code{data.frame} or \code{data.table} containing plot measurement data.}

\item{tree}{A \code{data.frame} or \code{data.table} containing tree information.}

\item{plot}{A \code{data.frame} or \code{data.table} containing plot information.}

\item{codesToExclude}{Character vector of damage agent codes used to filter tree data
(see GOA PSP Manual). Measurements with these codes will be removed.}

\item{excludeAllObs}{Logical. If \code{TRUE}, all prior and future observations of trees with
codes in \code{codesToExclude} are removed.}

\item{areaDiffThresh}{Numeric. Threshold for plot size discrepancy below which plots
are assigned a new ID. Expressed as \code{min(PlotSize)/max(PlotSize)}.}

\item{sppEquiv}{A table providing species name equivalencies between the original PSP species names
and the final standardized naming format. Default is \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{Character string. The column in \code{sppEquiv} that contains the standardized species names.
Default is \code{"Latin_full"}.}
}
\value{
A list containing standardized \code{plotData} and \code{treeData} as \code{data.table}s.
}
\description{
This function cleans and standardizes the Alberta PSP data, including tree measurements
and plot information. Species names can be standardized using a species equivalency table.
}
