% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlbertaPSP.R
\name{dataPurification_ABPSP}
\alias{dataPurification_ABPSP}
\title{standardize and treat the Alberta PSP data}
\usage{
dataPurification_ABPSP(
  treeMeasure,
  plotMeasure,
  tree,
  plot,
  codesToExclude = 3,
  excludeAllObs = TRUE,
  areaDiffThresh = 0.95,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{treeMeasure}{the tree measurement csv}

\item{plotMeasure}{the plot_measurement csv}

\item{tree}{the tree csv}

\item{plot}{the plot csv}

\item{codesToExclude}{damage agent codes used to filter tree data - see GOA PSP Manual.
Measurements with these codes will be removed}

\item{excludeAllObs}{if removing observations of individual trees due to damage codes,}

\item{areaDiffThresh}{the threshold of plot size discrepancy to allow below which
plots will be given a new ID column. Expressed as \code{min(PlotSize)/max(PlotSize)}
remove all prior and future observations if \code{TRUE}.}

\item{sppEquiv}{sdfsd}

\item{sppEquivCol}{sdfd}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the Alberta PSP data
}
