% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPSP.R
\name{dataPurification_BCPSP}
\alias{dataPurification_BCPSP}
\title{#' Standardize and Treat the BC PSP Data}
\usage{
dataPurification_BCPSP(
  treeDataRaw,
  plotHeaderDataRaw,
  damageAgentCodes,
  codesToExclude = "IBM",
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{treeDataRaw}{A \code{data.frame} or \code{data.table} containing the raw tree measurement data.}

\item{plotHeaderDataRaw}{A \code{data.frame} or \code{data.table} containing the raw plot header data.}

\item{damageAgentCodes}{A character vector of damage agent codes present in the data.}

\item{codesToExclude}{A character vector of damage agent codes to exclude from measurements.}

\item{excludeAllObs}{Logical. If \code{TRUE}, all prior and future observations of a tree with a damage code
in \code{codesToExclude} are removed. Default is \code{TRUE}.}

\item{sppEquiv}{A table providing species name equivalencies between the original PSP species names
and the final standardized naming format. Default is \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{Character string. The column in \code{sppEquiv} that contains the standardized species names.
Default is \code{"Latin_full"}.}
}
\value{
A list containing standardized \code{plotData} and \code{treeData} as \code{data.table}s.
}
\description{
This function cleans and standardizes the British Columbia PSP data, including tree measurements
and plot header information. Species names can be standardized using a species equivalency table.
}
