% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPSP.R
\name{dataPurification_BCPSP}
\alias{dataPurification_BCPSP}
\title{standardize and treat the BC PSP data}
\usage{
dataPurification_BCPSP(
  treeDataRaw,
  plotHeaderDataRaw,
  damageAgentCodes,
  codesToExclude = "IBM",
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA
)
}
\arguments{
\item{treeDataRaw}{the tree measurement csv}

\item{plotHeaderDataRaw}{the plot header csv}

\item{damageAgentCodes}{vector of damage agent codes}

\item{codesToExclude}{damage agents to exclude from measurements}

\item{excludeAllObs}{if removing observations of individual trees due to damage codes,
remove all prior and future observations if \code{TRUE}.}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the BC PSP data
}
