% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewBrunswickPSP.R
\name{dataPurification_NBPSP}
\alias{dataPurification_NBPSP}
\title{standardize and treat the New Brunswick PSP data}
\usage{
dataPurification_NBPSP(
  NB_PSP_Data,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{NB_PSP_Data}{list of data tables resulting from \code{prepInputsNBPSP}}

\item{sppEquiv}{sdfsd}

\item{sppEquivCol}{sdfd}
}
\value{
a list of standardized plot and tree data.tables
}
\description{
standardize and treat the New Brunswick PSP data
}
