% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewBrunswickPSP.R
\name{dataPurification_NBPSP}
\alias{dataPurification_NBPSP}
\title{standardize and treat the New Brunswick PSP data}
\usage{
dataPurification_NBPSP(
  PSP_PLOTS,
  PSP_TREE_YIMO,
  codesToExclude = NULL,
  excludeAllObs = TRUE
)
}
\arguments{
\item{PSP_PLOTS}{NB plot}

\item{PSP_TREE_YIMO}{NB tree}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the New Brunswick PSP data
}
