% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFIPSP.R
\name{dataPurification_NFIPSP}
\alias{dataPurification_NFIPSP}
\title{standardize and treat the NFI PSP data}
\usage{
dataPurification_NFIPSP(NFIdata, codesToExclude = "IB", excludeAllObs = TRUE)
}
\arguments{
\item{NFIdata}{list of NFI tree, plot, and location data}

\item{codesToExclude}{damage agents to exclude from measurements}

\item{excludeAllObs}{if removing observations of individual trees due to damage codes,
remove all prior and future observations if \code{TRUE}.}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the NFI PSP data
}
