% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFIPSP.R
\name{dataPurification_NFIPSP}
\alias{dataPurification_NFIPSP}
\title{standardize and treat the NFI PSP data}
\usage{
dataPurification_NFIPSP(lgptreeRaw, lgpHeaderRaw, approxLocation)
}
\arguments{
\item{lgptreeRaw}{the tree measurement data}

\item{lgpHeaderRaw}{the plot header data}

\item{approxLocation}{the location data}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the NFI PSP data
}
