% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuebecPSP.R
\name{dataPurification_QCPSP}
\alias{dataPurification_QCPSP}
\title{standardize and treat the QUEBEC PSP data}
\usage{
dataPurification_QCPSP(
  QuebecPSP,
  codesToExclude = NULL,
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "LandR"
)
}
\arguments{
\item{QuebecPSP}{list of PSP data.tables obtained via \code{prepInputsQCPSP}}

\item{codesToExclude}{TODO: eventually add codes for pest disturbance if applicable}

\item{excludeAllObs}{assuming codesToExclude is not NULL, exclude these obs or prior ones too}

\item{sppEquiv}{table of species names - see \code{LandR::sppEquiv}}
}
\value{
a list of standardized plot and tree data.tables
}
\description{
standardize and treat the QUEBEC PSP data
}
