% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuebecPSP.R
\name{dataPurification_QCPSP}
\alias{dataPurification_QCPSP}
\title{standardize and treat the QUEBEC PSP data}
\usage{
dataPurification_QCPSP(QuebecPSP, codesToExclude = NULL, excludeAllObs = TRUE)
}
\arguments{
\item{QuebecPSP}{list of PSP data.tables obtained via \code{prepInputsQCPSP}}
}
\value{
a list of standardized plot and tree data.tables
}
\description{
standardize and treat the QUEBEC PSP data
}
