% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaskatchewanPSP.R
\name{dataPurification_SKPSP}
\alias{dataPurification_SKPSP}
\title{Standardize and Treat the Saskatchewan PSP Data}
\usage{
dataPurification_SKPSP(
  SADataRaw,
  plotHeaderRaw,
  measureHeaderRaw,
  treeDataRaw,
  codesToExclude = NULL,
  excludeAllObs = TRUE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{SADataRaw}{A \code{data.frame} or \code{data.table} containing raw tree measurement data.}

\item{plotHeaderRaw}{A \code{data.frame} or \code{data.table} containing raw plot header data.}

\item{measureHeaderRaw}{A \code{data.frame} or \code{data.table} containing raw measurement header data.}

\item{treeDataRaw}{A \code{data.frame} or \code{data.table} containing tree data.}

\item{codesToExclude}{Character vector of damage agent codes used to filter tree data.
Codes: Natural or Undetermined = 1, Disease = 2, Insect = 3, Human = 4,
Wind = 5, Snow = 6, Other Trees = 7, Hail or Ice Storm = 8.
Measurements with these codes will be removed.}

\item{excludeAllObs}{Logical. If \code{TRUE}, all prior and future observations of trees with
codes in \code{codesToExclude} are removed.}

\item{sppEquiv}{A table providing species name equivalencies between the original PSP species names
and the final standardized naming format. Default is \code{LandR::sppEquivalencies_CA}.}

\item{sppEquivCol}{Character string. The column in \code{sppEquiv} that contains the standardized species names.
Default is \code{"Latin_full"}.}
}
\value{
A list containing standardized \code{plotData} and \code{treeData} as \code{data.table}s.
}
\description{
This function cleans and standardizes the Saskatchewan PSP data, including tree measurements,
plot headers, and measurement headers. Species names can be standardized using a species equivalency table.
}
