% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaskatchewanTSP.R
\name{dataPurification_SKTSP_Mistik}
\alias{dataPurification_SKTSP_Mistik}
\title{standardize and treat the Saskatchewan Mistik TSP data - I do not know where this comes from}
\usage{
dataPurification_SKTSP_Mistik(compiledPlotData, compiledTreeData)
}
\arguments{
\item{compiledPlotData}{the plot header data}

\item{compiledTreeData}{the tree data}
}
\value{
a list of plot and tree data.tables
}
\description{
standardize and treat the Saskatchewan Mistik TSP data - I do not know where this comes from
}
