% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R
\name{detect_dbh_outliers}
\alias{detect_dbh_outliers}
\title{-----------------------------------------------------------   detect_dbh_outliers function()  ----------------------------------------------
Detect and Flag DBH Outliers using Quantile Regression}
\usage{
detect_dbh_outliers(Trees, lower_tau = 0.05, upper_tau = 0.95)
}
\arguments{
\item{Trees}{A \code{data.table} or \code{data.frame} with columns: "DBH", "OrigPlotID1", "TreeNumber", "MeasureYear", etc.}

\item{lower_tau}{Lower quantile threshold (default = 0.05)}

\item{upper_tau}{Upper quantile threshold (default = 0.95)}
}
\value{
A list containing:
\describe{
\item{Trees}{The full data.table with \code{outlier_5th}, \code{outlier_95th}, and \code{outlier_type}.}
\item{OrigPlotID1s}{Unique \code{OrigPlotID1} identifiers.}
}
}
\description{
Flags implausible DBH values (static outliers) using quantile regression,
to help identify data issues in repeated tree measurements.
}
