% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSP.R
\name{getPSP}
\alias{getPSP}
\title{return a merged PSP object from a vector of data sources}
\usage{
getPSP(
  PSPdataTypes,
  destinationPath,
  forGMCS = FALSE,
  sppEquiv = LandR::sppEquivalencies_CA,
  sppEquivCol = "Latin_full"
)
}
\arguments{
\item{PSPdataTypes}{character vector of PSP data sources - e.g. \verb{c("BC", "SK)"}
Use \code{"all"} to get all available sources, and \code{"dummy"} for freely available data}

\item{destinationPath}{destination folder for downloaded objects}

\item{forGMCS}{if \code{TRUE}, will pre-filter plots with insect mortality to avoid
attributing insect mortality with climate}

\item{sppEquiv}{species equivalencies table.}
}
\value{
a list of standardized plot and tree data.tables
}
\description{
return a merged PSP object from a vector of data sources
}
