% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCPSP.R
\name{prepInputsBCPSP}
\alias{prepInputsBCPSP}
\title{source the BC PSP data}
\usage{
prepInputsBCPSP(dPath)
}
\arguments{
\item{dPath}{passed to prepInputs destinationPath}
}
\value{
a list of BC PSP objects
}
\description{
source the BC PSP data
}
