% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R
\name{treenum_to_multiplePSP}
\alias{treenum_to_multiplePSP}
\title{Identify Tree Numbers Linked to Multiple Species Within a Plot}
\usage{
treenum_to_multiplePSP(Trees)
}
\arguments{
\item{Trees}{A \code{data.table} of tree observations containing OrigPlotID1, Species,
PSP, MeasureYear and TreeNumber columns.}
}
\value{
A list containing:
\describe{
\item{Trees_corrected}{Cleaned tree dataset.}
}
}
\description{
Ensures consistent Species and Tree numbering within a plot (OrigPlotID1)
}
